Create security endpoints for Users

+++ src/main/java/com/example/travelproject/service/CommentsService.java
@@ -1,7 +1,10 @@
 package com.example.travelproject.service;
 
+import com.example.travelproject.domain.Attractions;
 import com.example.travelproject.domain.Comments;
+import com.example.travelproject.domain.Users;
 import com.example.travelproject.repository.CommentRepository;
+import com.example.travelproject.security.service.SecurityService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -12,9 +15,12 @@ import java.util.Optional;
 @Slf4j
 public class CommentsService {
     private final CommentRepository commentRepository;
+    private final SecurityService securityService;
 
-    public CommentsService(CommentRepository commentRepository) {
+    public CommentsService(CommentRepository commentRepository, SecurityService securityService) {
         this.commentRepository = commentRepository;
+
+        this.securityService = securityService;
     }
 
 
@@ -26,19 +32,24 @@ public class CommentsService {
         return commentRepository.findById(id);
     }
 
-    public Boolean createComments(Comments comments) {
-        try {
-            commentRepository.save(comments);
-            log.info(String.format("comments created " + comments.getText()));
-        } catch (Exception e) {
-            log.warn(String.format("error", comments.getText()));
-            return false;
-        }
-        return true;
+    public Comments addComments(String text, Users users, Attractions attractions) {
+        Comments comments = new Comments();
+        comments.setText(text);
+        comments.setUsers(users);
+        comments.setAttractions(attractions);
+        return commentRepository.save(comments);
     }
 
-    public void deleteCommentsById(Long id) {
-        commentRepository.deleteById(id);
+    public void deleteCommentsById(Long id, Long userId) {
+        try {
+            Comments comments = commentRepository.getReferenceById(id);
+            if (comments.getUsers().getId().equals(userId) || securityService.checkAccessById(userId)) {
+                commentRepository.deleteById(id);
+            }
+            log.info(String.format("comments delete id: " + comments.getId()));
+        } catch (Exception e) {
+            log.warn(String.format("error", id, e));
+        }
     }
 
     public Boolean updateComments(Comments comments) {
@@ -51,5 +62,4 @@ public class CommentsService {
         }
         return true;
     }
-
 }
